requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0
playwright==1.44.0
apscheduler==3.10.4
python-dotenv==1.0.1
jinja2==3.1.4
pytest==8.2.0
